requirements.txt
flask
selenium
python-dotenv
requests
